Add files via upload
This material is designed for beginners who want to learn development with a solid foundation. To become a good programmer, it’s important to understand not only coding, but also the fundamentals that support it. Every developer should have at least a basic idea about Operating Systems, Computer Architecture, and UML/Software Conception, because these elements form the backbone of how programs run, how they are designed, and how systems work.

In this lesson, we take the first step into Operating Systems, while continuing to strengthen essential skills like algorithms and data structures. Throughout the learning path, I will regularly refer to these core fields, because they are an important part of becoming a well-rounded and capable developer.